refactor: disable wayland in winit shell by default
